requirements.txt
python==3.8
pandas==2.2.3
seaborn==0.13.2
matplotlib==3.7.5
numpy==1.24.3
scikit-allel==1.3.8
imbalanced-learn==0.12.4
scikit-learn==1.3.2
keras==2.13.1
tensorflow==2.13.0
scikeras==0.11.0
xgboost==2.1.4
scipy==1.10.1
tqdm==4.67.1